Refactor: change Project package structure
