+++ src/main/java/subway/controller/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.controller.dto;
 
 public class StationResponse {
     private Long id;
